Added Zoom in, Zoom Out, Blur and reset functions
Created the Zoom function, the Blur function and the reset ability.
There is also a create thumbnail function in there, but it isn't mapped
to anything.

Made the code a little less hardcodey, so we don't need to worry about
how many images there are or even there names. We point at a directory
and pull all the jpgs out.

Removed the "q" button from closing the window. Intend on using it for
panning the image to the left, along with using "e" to pan to the right.

Mapped the Zoom In feature to the "w" button.
Mapped the Zoom Out feature to the "s" button
Mapped the Blur feature to the "b" button

Known bugs:
Due to the lack of proper limits in the code at the moment zooming in
too far will cause lots of lag.
Zooming out too far causes a crash.
Blur is only based on the original state of the image, not based on how
zoomed or blurred the image is currently.

I intend on creating a cache to preserve all the affects currently
applied.